cleaned up for final commit - renamed nanik_edition as LPD.py as it will be the main version of LPD (for now)
