+++ tests/lpd_digital_test.py
